Add PrimitiveTypes Long and ULong
The enum PrimitiveType represents, according to the comment, "the C++
built-in types.". However, it does not contain values for C++'s "long"
or "unsigned long", and the parser converts long to int. This is not
correct, and is a problem with 64bit support.

This patch adds Long and ULong values to PrimitiveType, and adds those
to all the switch-cases all around the code. For now Long and ULong are
handled exactly like Int and UInt, so this patch does not change
behavior.

+++ src/Parser/Parser.cpp
@@ -1341,8 +1341,8 @@ static CppSharp::AST::PrimitiveType WalkBuiltinType(const clang::BuiltinType* Bu
     case clang::BuiltinType::Int: return PrimitiveType::Int32;
     case clang::BuiltinType::UInt: return PrimitiveType::UInt32;
 
-    case clang::BuiltinType::Long: return PrimitiveType::Int32;
-    case clang::BuiltinType::ULong: return PrimitiveType::UInt32;
+    case clang::BuiltinType::Long: return PrimitiveType::Long;
+    case clang::BuiltinType::ULong: return PrimitiveType::ULong;
     
     case clang::BuiltinType::LongLong: return PrimitiveType::Int64;
     case clang::BuiltinType::ULongLong: return PrimitiveType::UInt64;
